Introduce a filesystem object
Ignore-this: 91e9a3de0f859e7844e295fa73178201

This patch removes the ODBStorage global variables, in favor of using a
filesystem object. We pick an object because we need to override various method
in the future inherited classes. The removal of the global variables in ODBStorage
should allow us to handle more than one repository in the CLI version of oasis-db.

We plan to implement a filesystem based on cache and remote sources for the CLI
(i.e. a filesystem that will use ODBSync data to list files and download them
as needed).

This also allow the creation of a git based filesystem, if needed.

darcs-hash:20110404082830-4d692-c7c0fed6b7093d27a932e7357fc6f148b0bdc41d